requirements.txt
FastAPI==0.70.0
SQLAlchemy==1.4.22
requests==2.26.0
python-dotenv==0.19.0
pytest==6.2.5
bandit==1.7.0
starlette==0.17.1
uvicorn==0.15.0
slowapi==0.1.4
python-jose==3.3.0
passlib[bcrypt]==1.7.4
webdriver-manager==3.5.2
selenium==3.141.0
prometheus-fastapi-instrumentator==2.5.0
